Refs #31. Rename LogPickWindow.
